requirements.txt
streamlit
langchain_groq
langchain_community
langchain_core
langchain_huggingface
langchain
huggingface-hub
FAISS
python-dotenv
reportlab
pymupdf
pdf2image